[SPARK-46755][K8S][INFRA][DOCS] Upgrade `Volcano` to 1.8.2
### What changes were proposed in this pull request?

This PR aims to upgrade `Volcano` to 1.8.1 in K8s integration test document and GitHub Action job.

### Why are the changes needed?

To bring the latest bug fixes in addition to the test coverage for Volcano scheduler 1.8.2.
- https://github.com/volcano-sh/volcano/releases/tag/v1.8.2

Note that the following `Volcano` bug fix reveals our test suite issue. So, it's disabled in Volcano testing.
- https://github.com/volcano-sh/volcano/issues/3271

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #44779 from dongjoon-hyun/SPARK-46755.

Authored-by: Dongjoon Hyun <dhyun@apple.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>