[SPARK-46819][CORE] Move error categories and states into JSON
### What changes were proposed in this pull request?

Move the error categories and states out of the Markdown table in the error README and into standalone JSON files.

I am using the terms "error category" and "error state" as defined in SPARK-46810.

### Why are the changes needed?

The data captured in Markdown is not automation-friendly. In future work, I intend to automate the generation of [sql-error-conditions-sqlstates.md][states] using these new JSON files.

Additionally, there are many instances of duplicate keys and inconsistent descriptions for the error categories, which this change resolves by moving the data into JSON. This also enables IDE support for basic validation.

[states]: https://github.com/apache/spark/blob/561b995c69d9405ae900fbca0c9e6b63b9449a14/docs/sql-error-conditions-sqlstates.md

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

The updated unit tests ensure the following:
- All error categories (e.g. `42`) and states (e.g. `42K01`) have the correct length and format.
- All error states have a two-character prefix that is a valid error category.
- All error classes define an error state that is valid.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #44863 from nchammas/SPARK-46819-error-data-automation.

Authored-by: Nicholas Chammas <nicholas.chammas@gmail.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>