[SPARK-46888][CORE] Fix `Master` to reject `/workers/kill/` requests if decommission is disabled
### What changes were proposed in this pull request?

This PR aims to fix `Master` to reject `/workers/kill/` request if `spark.decommission.enabled` is `false` in order to fix the dangling worker issue.

### Why are the changes needed?

Currently, `spark.decommission.enabled` is `false` by default. So, when a user asks to decommission, only Master marked it `DECOMMISSIONED` while the worker is alive.
```
$ curl -XPOST http://localhost:8080/workers/kill/\?host\=127.0.0.1
```

**Master UI**
![Screenshot 2024-01-27 at 6 19 18 PM](https://github.com/apache/spark/assets/9700541/443bfc32-b924-438a-8bf6-c64b9afbc4be)

**Worker Log**
```
24/01/27 18:18:06 WARN Worker: Receive decommission request, but decommission feature is disabled.
```

To be consistent with the existing `Worker` behavior which ignores the request.

https://github.com/apache/spark/blob/dc0dff681ba60aaa3dcca8ef603e06750d948a3f/core/src/main/scala/org/apache/spark/deploy/worker/Worker.scala#L859-L868

### Does this PR introduce _any_ user-facing change?

No, this is a bug fix.

### How was this patch tested?

Pass the CI with the newly added test case.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #44915 from dongjoon-hyun/SPARK-46888.

Authored-by: Dongjoon Hyun <dhyun@apple.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>