[SPARK-46930][SQL] Add support for a custom prefix for Union type fields in Avro
### What changes were proposed in this pull request?

This PR enhances stable ids functionality in Avro by allowing users to configure a custom prefix for Union type member fields when `enableStableIdentifiersForUnionType` is enabled.

Without the patch, the fields are generated with `member_` prefix, e.g. `member_int`, `member_string`. This could become difficult to change for complex schemas.

The solution is to add a new option `stableIdentifierPrefixForUnionType` which defaults to `member_` and allows users to configure whatever prefix they require, e.g. `member`, `tmp_`, or even an empty string.

### Why are the changes needed?

Allows to customise the prefix of stable ids in Avro without the need to rename all of the columns which could be cumbersome for complex schemas.

### Does this PR introduce _any_ user-facing change?

Yes. The PR adds a new option in Avro: `stableIdentifierPrefixForUnionType`.

### How was this patch tested?

Existing tests + a new unit test to verify different prefixes.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #44964 from sadikovi/SPARK-46930.

Authored-by: Ivan Sadikov <ivan.sadikov@databricks.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>