[SPARK-46512][CORE] Optimize shuffle reading when both sort and combine are used
### What changes were proposed in this pull request?

After the shuffle reader obtains the block, it will first perform a combine operation, and then perform a sort operation. It is known that both combine and sort may generate temporary files, so the performance may be poor when both sort and combine are used. In fact, combine operations can be performed during the sort process, and we can avoid the combine spill file.
See  https://issues.apache.org/jira/browse/SPARK-46512 for details

### Why are the changes needed?

Reduce the number of spills to disk when both sort and combine are used

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

spark examples run in local and yarn mode.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #44512 from zhengchenyu/SPARK-46512.

Authored-by: zhengchenyu <zhengchenyu16@163.com>
Signed-off-by: Mridul Muralidharan <mridul<at>gmail.com>