[SPARK-47003][K8S] Detect and fail on invalid volume sizes (< 1KiB) in K8s
### What changes were proposed in this pull request?

This PR aims to detect and fails on invalid volume size.

### Why are the changes needed?

This happens when the user forget the unit of volume size. For example, `100` instead of `100Gi`.

### Does this PR introduce _any_ user-facing change?

For K8s volumes, the system is trying to use the system default minimum volume size. However it totally depends on the underlying system. And, this misconfiguration misleads the users in many cases because the job is started and running in unhealthy status.
- First, the executor pods will be killed by the K8s control plane due to the out of disk situation.
- Second, Spark is trying to create new executors (still with small disks) and to retry multiple times.

We had better detect the missed-unit situation and make those jobs fail as early as possible.

### How was this patch tested?

Pass the CIs with newly added test cases.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #45061 from dongjoon-hyun/SPARK-47003.

Authored-by: Dongjoon Hyun <dhyun@apple.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>