[SPARK-46865][SS] Add Batch Support for TransformWithState Operator
### What changes were proposed in this pull request?

We are allowing batch queries to use and define the `TransformWithState` operator, which was initially introduced for streaming.

### Why are the changes needed?

This is needed to keep up the parity between streaming and batch APIs, since we want everything supported in streaming to be supported in batch, as well.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Added unit tests that use the TransformWithState operator with a batch query.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #44884 from ericm-db/tws-batch.

Lead-authored-by: Eric Marnadi <eric.marnadi@databricks.com>
Co-authored-by: ericm-db <132308037+ericm-db@users.noreply.github.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>