[SPARK-47071][SQL] Inline With expression if it contains special expression
### What changes were proposed in this pull request?

This is a bug fix for the With expression. If the common expression contains special expression like aggregate expresson, we cannot pull it out and put it in Project. We have to inline it.

### Why are the changes needed?

bug fix.

### Does this PR introduce _any_ user-facing change?

a failed can run after this fix, but this bug is not released yet.

### How was this patch tested?

new tests

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #45134 from cloud-fan/with.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>