[MINOR][DOCS] Show sort order of NaN relative to infinity
### What changes were proposed in this pull request?

Tweak the examples in our documentation on NaN semantics to demonstrate how NaN sorts relative to infinity.

### Why are the changes needed?

The docs make an interesting point about how NaNs are sorted, but none of the examples that follow demonstrate this point.

https://github.com/apache/spark/blob/a6d5b6a2f89fb8aa8de835e7bf5c72b1bed32bf7/docs/sql-ref-datatypes.md#L294

### Does this PR introduce _any_ user-facing change?

Yes, it tweaks some documentation.

### How was this patch tested?

I ran the SQL I tweaked to confirm the output.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #45047 from nchammas/nan-sorts-after-all.

Authored-by: Nicholas Chammas <nicholas.chammas@gmail.com>
Signed-off-by: Peter Toth <peter.toth@gmail.com>