[SPARK-47132][DOCS][PYTHON] Correct docstring for pyspark's dataframe.head
### What changes were proposed in this pull request?

Change description of `returns` in docstring for pyspark's dataframe's `head` to match actual behaviour.

### Why are the changes needed?

The docstring claimed that `head(n)` would return a `Row` (rather than a list of rows) iff n == 1, but that's incorrect.

Type hints, example, and implementation show that the difference between row or list of rows lies in whether n is supplied at all -- if it isn't, `head()` returns a `Row`, if it is, even if it is 1, `head(n)` returns a list.

### Does this PR introduce _any_ user-facing change?

No, only a Python docstring is changed.

### How was this patch tested?

No tests were added since only docs changed.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #45197 from wunderalbert/patch-1.

Authored-by: Albert Ziegler <55346556+wunderalbert@users.noreply.github.com>
Signed-off-by: Xinrong Meng <xinrong@apache.org>