[SPARK-47159][INFRA] Set `OBJC_DISABLE_INITIALIZE_FORK_SAFETY=YES` in `MacOS` GitHub Action Job
### What changes were proposed in this pull request?

This PR aims to set `OBJC_DISABLE_INITIALIZE_FORK_SAFETY=YES` in `MacOS` GitHub Action Job.
- https://github.com/apache/spark/actions/workflows/build_maven_java21_macos14.yml

### Why are the changes needed?

This is recommended for Python tests, but it seems that `Spark Connect` Python test suites are also related. We had better have this to be safe or isolate any issue.

https://github.com/apache/spark/blob/258d880a4a19104e3a79bc889038eccf24e1a76e/python/docs/source/development/testing.rst?plain=1#L39

### Does this PR introduce _any_ user-facing change?

No, this is a test infra-only change.

### How was this patch tested?

Manual review. This is tested by Daily CI after merging.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #45249 from dongjoon-hyun/SPARK-47159.

Authored-by: Dongjoon Hyun <dhyun@apple.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>