[SPARK-47268][SQL][COLLATIONS] Support for repartition with collations
### What changes were proposed in this pull request?

With this change we add support for spark partitioning (and inherently shuffle) for columns that are collated with non-binary collations.

For overview of collation effort please refer to the parent JIRA ticket.

### Why are the changes needed?

Without this change, partitioning/shuffle operator will not respect collation of `StringType`. In other words, rows that are considered to be equal under provided collation can end up in different partitions.

### Does this PR introduce _any_ user-facing change?

Yes - PR fixes partitioning behaviour.

### How was this patch tested?

Additional test is added to collation suite.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #45370 from dbatomic/shuffle_and_collations.

Authored-by: Aleksandar Tomic <aleksandar.tomic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>