[SPARK-39771][CORE] Add a warning msg in `Dependency` when a too large number of shuffle blocks is to be created
### What changes were proposed in this pull request?
Add a warning msg in `Dependency` when a large number of shuffle blocks is to be created which may crash the driver with OOM.

### Why are the changes needed?
Warn user ahead of a potential failure, and hint a possible solution.

### Does this PR introduce _any_ user-facing change?

User may see a warning message as above when the number of shuffle blocks exceeds our threshold. There is no change to the execution outcome.

### How was this patch tested?

I added an unit test to ensure the warning message will be logged.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #45266 from y-wei/y-wei/SPARK-39771.

Lead-authored-by: y-wei <yifei.wei@databricks.com>
Co-authored-by: Yifei Wei <52417396+y-wei@users.noreply.github.com>
Signed-off-by: Yuanjian Li <yuanjian.li@databricks.com>