[SPARK-47300][SQL] `quoteIfNeeded` should quote identifier starts with digits
### What changes were proposed in this pull request?

`quoteIfNeeded` is used to generate pretty strings of identifiers in error message, EXPLAIN result, etc. It's mostly for humans to read but sometimes people may copy-paste the identifier string and put it in a SQL. There is a small issue in `quoteIfNeeded`: it does not quote number literals like `0d`, and people will get parse exception if they directly put the identifier string in SQL.

This PR fixes the issue by always quoting the identifier if it does not start with alphabet or underscore. Note, there might be program trying to parse the identifier string, but this change is safe as these programs should already handle quoting and it's ok to quote more.

### Why are the changes needed?

make identifier string parsable is more user-friendly.

### Does this PR introduce _any_ user-facing change?

yes, the identifier string in the error message and EXPLAIN result will be properly quoted.

### How was this patch tested?

new test suite

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #45401 from cloud-fan/quote.

Lead-authored-by: Wenchen Fan <wenchen@databricks.com>
Co-authored-by: Wenchen Fan <cloud0fan@gmail.com>
Signed-off-by: Kent Yao <yao@apache.org>