[SPARK-47218][SQL] XML: Changed SchemaOfXml to fail on DROPMALFORMED mode
### What changes were proposed in this pull request?

Changed schema_of_xml should fail with an error on DROPMALFORMED mode to avoid creating schemas out of invalid XML.

### Why are the changes needed?

DROPMALFORMED parse mode imply silently dropping the malformed record. But SchemaOfXml is expected to return a schema and may not have a valid schema to return for a malformed record. So DROPMALFORMED cannot be supported..

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Unit test.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #45379 from yhosny/xml-parsemode-error.

Authored-by: Yousof Hosny <yousof.hosny@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>