[SPARK-47341][CONNECT] Replace commands with relations in a few tests in SparkConnectClientSuite
### What changes were proposed in this pull request?

A few [tests](https://github.com/apache/spark/blob/master/connector/connect/client/jvm/src/test/scala/org/apache/spark/sql/connect/client/SparkConnectClientSuite.scala#L481-L527) in SparkConnectClientSuite attempt to test the result collection of a reattachable execution through the use of a SQL command. The SQL command, on a real server, is not executed eagerly (since it is a select command) and thus, is not entirely accurate. The test itself is non-problematic since a dummy server with dummy responses is used but a small improvement here would be to construct a relation rather than a command.

### Why are the changes needed?

Although these tests are not problematic, we should ensure that the behavior of the tests is consistent with the actual behavior of real servers to avoid misleading developers.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

The tests in org.apache.spark.sql.connect.client.SparkConnectClientSuite passed.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #45460 from xi-db/SPARK-47341-fix-misleading-tests.

Authored-by: Xi Lyu <xi.lyu@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>