[SPARK-47343][SQL] Fix NPE when `sqlString` variable value is null string in execute immediate
### What changes were proposed in this pull request?
In this PR I propose a change to add new error class NULL_QUERY_STRING_EXECUTE_IMMEDIATE that is thrown when sqlString variable is a null string in execute immediate query. Null pointer exception was thrown before this change.

### Why are the changes needed?
To ensure that user facing error is as clear as possible and to avoid NPE for possible user input.

### Does this PR introduce _any_ user-facing change?
Yes. This PR introduces a new error class that has an user facing description.

### How was this patch tested?
By running an "EXEC IMMEDIATE - Null string as sqlString parameter" test from AnalysisErrorSuite.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #45462 from stefanbuk-db/ES-1055420_fix-ExecuteImmediateQuery-sqlString-null.

Authored-by: Stefan Bukorovic <stefan.bukorovic@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>